Backport http keep-alive configuration to enable EventSource reconnects
Node.js v19 changed the default configuration of the built-in http agent used by EventSource to connect to Server Sent Events from ESPHome.  Agents before this change did not use sockets with keep-alive enabled, preventing `EventSource` from detecting disruptions to the connection it uses to receive events.

This change will detect if the node version is old enough not to have this behavior and configure the http (and https) modules to match behavior from more recent Node.js versions.